MANY CHANGES and a partial re-implementation of a number of classes.
The event project has been moved to the DJUTILS project and can be found as part of djutils. See https://djutils.org/manual/djutils-project/event-package for more information. Much of the event package has been re-implemented and streamlined as part of the the transfer to DJUTILS.  

As a result of the transfer of the event project, all references to nl.tudelft.simulation.dsol.event.* had to be changed to org.djutils.event.*

The DJUNITS event project implements the EventProducer as an abstract class that needs a sourceId for firing events rather than exposing 'this' in each event. Therefore, the constructor of the classes that extend EventProducer had to be changed. 

The DJUNITS event project forces the sourceId and content of an event, and the timestamp of a TimedEvent to be Serializable, because events are often streamed into a database. Several classes had to be updated to reflect this change.  

The context package in the dsol-naming project has been completely re-implemented. The dependency on underlying  JNDI classes has been reduced, and the implementation is more lightweight than it was originally. The package  structure has been updated, and several of the classes have moved to new packages. Interfaces with a Name object have been removed; only the String representations have been re-implemented. Many of the other interfaces of the new context implementation are the same as or similar to the original ones. The new Context implementation has been documented in https://simulation.tudelft.nl/dsol/manual/advanced/naming-context.

Rigid testing of the context package revealed many smaller and bigger mistakes (one of the reasons for re-implementing the entire package).

Unit tests have been included for the entire context package and dsol-naming project.

A context demo has been added to the dsol-naming project under the test tree (DemoServer and DemoClient classes).

The CategoryLogger in DJUTILS has been turned into a singleton rather than a static class. SimLogger implements a logger with functionality of the CategoryLogger, but the logger is simulator-aware and can print the simulator time as part of the log message.