Fixed a bug where the run thread of the simulator / animator does not start very briefly after it has been stopped, since it is possible that the worker thread has not yet stopped while the stop() method has already returned.
Fixed the way how experiments and replications are triggered and advanced. The old way was to start at experiment number -1, and replication number -1, and trigger an END_OF_EXPERIMENT_EVENT and an END_OF_REPLICATION_EVENT at the start to advance the experiment and replication both to 0. This has been replaced by a proper start that also triggers the right events.   

Events for starting and stopping the simulation have been split into a STARTING_EVENT (Event) that indicates that an external thread has asked to start the simulator, and a START_EVENT (TimedEvent) that indicates that the run() method of the simulation has actually started. Same for the STOPPING_EVENT (Event) and the STOP_EVENT (TimedEvent).    

The way events are fired from the Simulator has been made much more precise. Start, stop, end_replication and other events are fired as close as possible to the actual state change that happens. Double firing of events is prevented as much as possible.      

The methods for the simulator start(boolean), step(boolean) and stop(boolean) have been removed. These methods could indicate not to fire a START, STEP or STOP event. Since sending these events is part of the contract for a simulator, the option to not send them has been removed. Instead, startImpl(), stepImpl() and stopImpl() have been created that just do the work without firing events that belong to the user interface thread.

The START_REPLICATION_EVENT, WARMUP_EVENT and END_REPLICATION_EVENT are now defined in the Replication   rather than in the SimulatorInterface.

Fixed the BoxAndWhiskerPlot for not listening to TIMED event for timed, weighted tally.