Took out the STEP_EVENT. Instead, the step follows the same rule as normal running: the START_EVENT and the STOP_EVENT are fired before/after the execution of the step.
More clear contract for the old isRunning() method. Created a RunState class and a method getRunState() that is much more precise in the state of the simulator: NOT_INITIALIZED, INITIALIZED, STARTING, STARTED, STOPPING, STOPPED, ENDED with more precise methods: isStartingOrRunning(),  isStoppingOrStopped(). The Simulator now implements a clear state machine for the run state.

Created a clear contract for the Replication state. Created a ReplicationState class and a method getReplicationState() that is precise in the state of the replication within simulator: NOT_INITIALIZED, INITIALIZED, STARTED, ENDING, ENDED. The Simulator implements a clear state machine for the replication state.