A RunControl object to more easily store offline data for the RunControl has been created. These objects can now be passed to the constructor of experiments and replications. Internally, these objects, rather than their individual fields, are used to store the information.
The stream management of the DSOLModel interface and AbstractDSOLModel have been streamlined. Within the model, the information about streams is stored in a StreamInfofrmation object rather than in an individual map. This abstracts the implementation from the function of the streams. 

A method getDefaultStream() has been added to the model to make it easier to use the streams in easy cases.