The dsol-animation-gis project has been completely reorganized. The GisFileXMLParser has a new location and is renamed to EsriFileXmlParser. Generic GIS interfaces have been moved to the dsol-animation project, so the dependency between dsol-swing and dsol-animation-gis is  not needed anymore. The actual implementation of the shape file parser and the osm parser have been moved to two new projects: dsol-animation-gis-esri and dsol-animation-gis-osm. This means that when dsol-swing is loaded, no osmosis or esri dependencies are used, and the osmosis libraries are only included when an OSM file has to be parsed.
Code for the ESRI shape parser has been rewritten, with a much lower memory footprint.

Map display in a new dsol-animation-gis-osm project can now also parse and display OpenStreetMap files (.pbf, .osm, .osm.gz and .osm.bz2).

Extra parsers for GIS files have been added to the dsol-animation-gis project that work on the basis of a much simpler CSV or TSV file rather than the verbose XML file to specify what has to be read. The XML file still works for upward compatibility (but is already for a long time not dependent on the relatively heavy Xerces library anymore).

Demos for the XML parsers have beem added to the dsol-demo project, in the n.t.s.dsol.animation.gis package.