The animation has been decoupled from dsol-core, and can now be used to display maps or animations without  having a DSOL simulator present. The animation package is still dependent on dsol-naming, which provides the (remote, file or memory) context in which the animation objects are registered and possibly transferred.
Because of the decoupling of animation from simulation, the Renderable2D classes for simulation objects now have to specify the simulator.getReplication() as the Context. Either they can do so directly, or these classes can extend the SimRenderable2D instead, which takes a simulator as the parameter to indicate where to  find the context.